v3.0: add resend spi loop if event was missed
This should improve stability when sending
a new message and if the response was not in the timeout we send it again
for up to ten times.
this has changed the timeout to 32milisec from 200mili sec but we loop
write so the total time is now up to 320milisec.